+++ crates/cfxcore/core/src/transaction_pool/deferred_pool/mod.rs
@@ -17,7 +17,6 @@ use std::{collections::HashMap, sync::Arc};
 #[cfg(test)]
 mod tests;
 
-
 /// The `DeferredPool` is designed to organize transactions for each address
 /// based on their nonce. It efficiently maintains and queries transactions even
 /// when received nonces are non-sequential. In addition, it calculates
